:recycle: refactor(#168/ai): ai 서비스, dto 관련 클래스 ninedot-application 모듈로
