+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/ai/dto/request/SubGoalRecommendationRequest.java
