+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/ai/dto/response/SubGoalRecommendationResponse.java
